// src/interfaces/EditForm.ts
export interface EditForm {
  _id: any;
  email: string | undefined;
  password: string | undefined;
  about: string | undefined;
  address: string | undefined;
  title: string | undefined;
  role: string | undefined;
  salary: number | undefined;
  occupancy: string | undefined;
  company: string | undefined;
}